msm: audio: qdsp6v2: Modify bytes consumed calculation
Modify bytes consumed calculation to avoid precision
loss


Conflicts:
	arch/arm/mach-msm/qdsp6v2/q6asm.c

+++ arch/arm/mach-msm/include/mach/qdsp6v2/q6asm.h
@@ -129,7 +129,7 @@ struct audio_client {
 	app_cb			cb;
 	void			*priv;
 	uint32_t         io_mode;
-	uint32_t         time_stamp;
+	uint64_t         time_stamp;
 };
 
 void q6asm_audio_client_free(struct audio_client *ac);
@@ -238,7 +238,7 @@ int q6asm_set_lrgain(struct audio_client *ac, int left_gain, int right_gain);
 /* Enable Mute/unmute flag */
 int q6asm_set_mute(struct audio_client *ac, int muteflag);
 
-uint32_t q6asm_get_session_time(struct audio_client *ac);
+uint64_t q6asm_get_session_time(struct audio_client *ac);
 
 /* Client can set the IO mode to either AIO/SIO mode */
 int q6asm_set_io_mode(struct audio_client *ac, uint32_t mode);
